--- a/vllm/model_executor/models/plamo2.py
+++ b/vllm/model_executor/models/plamo2.py
@@ -332,7 +332,7 @@ class Plamo2MambaMixer(MambaBase, CustomOp):
                                          dim=0)
             # Split along batch dimension
             state_indices_tensor_d, state_indices_tensor_p = torch.split(
-                state_indices_tensor[:num_actual_tokens],
+                state_indices_tensor,
                 [num_decodes, num_prefills],
                 dim=0,
             )