## Purpose

This PR follows the great work by @heheda12345 and @tdoublep to finally enable the v1 engine and full CUDA graphs (decode only) for the Plamo2 model architecture.
It also incorporates other recent improvements to `MambaMixer2` (such as #21075 and #18218).

Additionally, this PR fixes #22999 and consequently re-enables Plamo2 in the hybrid model unit tests. I also enabled Plamo2 for the full CUDA graph unit tests.

Note: To support (piecewise) CUDA graphs and `torch.compile`, the `Plamo2MambaMixer` is added to `CompilationConfig._attention_ops` by default. I think it's OK to do since the diff is minimal.

This PR is potentially in conflict with #21467 (cc @cyang49).

Fixes #23956.
Incorporates #23520 to fix #22999.

## Benchmarks

### Latency

Latency improves quite a bit.

This PR (v1 engine, `CUDAGraphMode.FULL_AND_PIECEWISE`):
```
$ VLLM_USE_V1=1 vllm bench latency --model pfnet/plamo-2-8b --trust-remote-code --max-model-len 8192 --no-enable-prefix-caching --input-len 1000 --output-len 1000 --num-iters 3 --num-iters-warmup 3

Avg latency: 26.398034774387877 seconds
10% percentile latency: 26.389116884209216 seconds
25% percentile latency: 26.39132913807407 seconds
50% percentile latency: 26.395016227848828 seconds
75% percentile latency: 26.403231137432158 seconds
90% percentile latency: 26.408160083182157 seconds
99% percentile latency: 26.411117450632155 seconds
```

`main` (v0 engine, piecewise):
```
Avg latency: 30.82136501589169 seconds
10% percentile latency: 30.770767986914144 seconds
25% percentile latency: 30.776328965090215 seconds
50% percentile latency: 30.781959515297785 seconds
75% percentile latency: 30.786822279333137 seconds
90% percentile latency: 30.853961815405636 seconds
99% percentile latency: 31.446262379428372 seconds
```

### Throughput

Throughput does not significantly change.

This PR (v1 engine, `CUDAGraphMode.FULL_AND_PIECEWISE`):
```
$ VLLM_USE_V1=1 vllm bench throughput --model pfnet/plamo-2-8b --trust-remote-code --max-model-len 8192 --no-enable-prefix-caching --dataset-name sharegpt --dataset-path /datasets/ShareGPT_V3_unfiltered_cleaned_split.json --num-prompts 1000

Throughput: 12.00 requests/s, 4619.67 total tokens/s, 2144.02 output tokens/s
```

`main` (v0 engine, piecewise):
```
Throughput: 11.83 requests/s, 4532.89 total tokens/s, 2104.20 output tokens/s
```

### Output quality

No output degradation.

This PR (v1 engine):
```
$ VLLM_USE_V1=1 lm_eval --model vllm  --model_args pretrained=pfnet/plamo-2-8b,tensor_parallel_size=1,dtype=auto,gpu_memory_utilization=0.95,enable_prefix_caching=False --batch_size auto --trust_remote_code  --cache_requests true --tasks gsm8k

|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.5163|±  |0.0138|
|     |       |strict-match    |     5|exact_match|↑  |0.6603|±  |0.0130|
```

`main` (v0 engine, piecewise):
```
$ VLLM_USE_V1=0 lm_eval --model vllm  --model_args pretrained=pfnet/plamo-2-8b,tensor_parallel_size=1,dtype=auto,gpu_memory_utilization=0.95,enable_prefix_caching=False --batch_size auto --trust_remote_code  --cache_requests true --tasks gsm8k

|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.5125|±  |0.0138|
|     |       |strict-match    |     5|exact_match|↑  |0.6581|±  |0.0131|
```

## Test Result

`pytest -s -v tests/models/language/generation/test_hybrid.py` passed.

---
<details>
<summary> Essential Elements of an Effective PR Description Checklist </summary>

- [x] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [x] The test plan, such as providing test command.
- [x] The test results, such as pasting the results comparison before and after, or e2e results
- [x] (Optional) The necessary documentation update, such as updating `supported_models.md` and `examples` for a new model.
- [ ] (Optional) Release notes update. If your change is user facing, please update the release notes draft in the [Google Doc](https://docs.google.com/document/d/1YyVqrgX4gHTtrstbq8oWUImOyPCKSGnJ7xtTpmXzlRs/edit?tab=t.0).
</details>